Move ruleset settings to own section
