+++ osu.Game/Localisation/RulesetSettingsStrings.cs
@@ -1,0 +1,19 @@
+// Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the MIT Licence.
+// See the LICENCE file in the repository root for full licence text.
+
+using osu.Framework.Localisation;
+
+namespace osu.Game.Localisation
+{
+    public static class RulesetSettingsStrings
+    {
+        private const string prefix = @"osu.Game.Resources.Localisation.RulesetSettings";
+
+        /// <summary>
+        /// "Rulesets"
+        /// </summary>
+        public static LocalisableString Rulesets => new TranslatableString(getKey(@"rulesets"), @"Rulesets");
+
+        private static string getKey(string key) => $@"{prefix}:{key}";
+    }
+}
