+++ osu.Game/Overlays/Settings/Sections/GameplaySection.cs
@@ -1,16 +1,11 @@
 ﻿// Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the MIT Licence.
 // See the LICENCE file in the repository root for full licence text.
 
-using System;
-using osu.Framework.Allocation;
 using osu.Framework.Graphics;
-using osu.Game.Overlays.Settings.Sections.Gameplay;
-using osu.Game.Rulesets;
-using System.Linq;
 using osu.Framework.Graphics.Sprites;
-using osu.Framework.Logging;
 using osu.Framework.Localisation;
 using osu.Game.Localisation;
+using osu.Game.Overlays.Settings.Sections.Gameplay;
 
 namespace osu.Game.Overlays.Settings.Sections
 {
@@ -34,25 +29,6 @@ namespace osu.Game.Overlays.Settings.Sections
                 new InputSettings(),
                 new ModsSettings(),
             };
-        }
-
-        [BackgroundDependencyLoader]
-        private void load(RulesetStore rulesets)
-        {
-            foreach (Ruleset ruleset in rulesets.AvailableRulesets.Select(info => info.CreateInstance()))
-            {
-                try
-                {
-                    SettingsSubsection section = ruleset.CreateSettings();
-
-                    if (section != null)
-                        Add(section);
-                }
-                catch (Exception e)
-                {
-                    Logger.Error(e, "Failed to load ruleset settings");
-                }
-            }
         }
     }
 }
